requirements.txt
Django>=3.0.0
djangorestframework>=3.10.3
drf-yasg>=1.17.0
uwsgi>=2.0.18
pytest>=5.3.2
cryptography>=2.8